Add files display

+++ app/model/ServerFile.scala
@@ -1,6 +1,6 @@
 package model
 
-class ServerFile(val name: String, val size: Int) extends ServerEntity {
+class ServerFile(val name: String, val size: Long) extends ServerEntity {
 
   val icon: String = "glyphicon-file"
   val content: List[ServerEntity] = Nil
