+++ app/model/ServerFolder.scala
@@ -3,9 +3,9 @@ package model
 class ServerFolder(val name: String, val content: List[ServerEntity]) extends ServerEntity {
 
   val icon: String = "glyphicon-folder-close"
-  val size: Int = sizeOf(content)
+  val size: Long = sizeOf(content)
 
-  def sizeOf(theContent: List[ServerEntity]): Int = theContent match {
+  def sizeOf(theContent: List[ServerEntity]): Long = theContent match {
     case (x::xs) => x.size + sizeOf( xs )
     case Nil => 0
   }
